interactive fish  working with LinearBlendSkinning on GPU
